move text confirm modal
