+++ packages/ui-patterns/Dialogs/TextConfirmModal.tsx
